moved configuration block model to reconfiguration directory
