+++ src/Simulator/Engine/Simulation.py
@@ -1,7 +1,8 @@
 from Parameters import Parameters
 
 from Chain.Node import Node
-from Chain.Block import Block, ConfigurationBlock
+from Chain.Block import Block
+from Chain.Reconfiguration.ConfigurationBlock import ConfigurationBlock
 from Chain.TransactionFactory import TransactionFactory
 
 from Utils import Tools
